requirements.txt
Django==1.11.9
django-crispy-forms==1.7.0
django-model-utils==3.1.1
djangorestframework==3.7.7